+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Missing global `SetName` function
+- `SetArgs` to set command line arguments manually
 
 
 ## [0.2.0] - 2017-12-19